+++ .gitignore
@@ -161,4 +161,4 @@ cython_debug/
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 #.idea/
 
-data_sync_microservice/config.json
+sigeca_data_export_microservice/config.json
