+++ README.md
@@ -1,5 +1,5 @@
 # sigeca-synchronization
 OpenLMIS Client used for synchronizing data with SIGECA Central Data Warehouse 
 
-# Data Synchronization Microservice 
-[README.md](./data_sync_microservice/README.md)
+# Sigeca Data Export Synchronization Microservice 
+[README.md](./sigeca_data_export_microservice/README.md)
