--- data_sync_microservice/app/application/services/__init__.py
@@ -1,1 +1,0 @@
-from .data_sync_service import DataSyncService
