+++ sigeca_data_export_microservice/app/application/services/__init__.py
@@ -1,0 +1,1 @@
+from .sigeca_data_export_service import DataSyncService
