+++ sigeca_data_export_microservice/app/application/synchronizations/resources_synchronization.py
