+++ sigeca_data_export_microservice/app/domain/models/sync_log.py
