+++ sigeca_data_export_microservice/app/domain/resources/abstract.py
