+++ sigeca_data_export_microservice/app/infrastructure/utils.py
